+++ CHANGELOG.md
@@ -32,6 +32,8 @@ Date format: (YYYY-MM-DD)
   * Internal: Moved ArgumentFilter into base commands lib package.
   * Internal: CommandArguments keep track of their parent argument now (if used internally by another argument) and use that for their error messages.
   * Internal: Added AliasArgument, that can be used to change the display name of an command argument (used in the command format). This may for example be useful in the presence of multiple, otherwise conflicting literal arguments.
+  * Internal: Minor changes to handling errors during command handling. Besides the stack trace, the plugin also logs the command context (parsed arguments) now.
+  * Internal: Added map view and toString to CommandContext.
 * Fixed: The shop creation item can no longer be used from dispensers if regular use is disabled.
 
 ## v2.8.1 (2019-08-23)